This code should allow you to create a dataframe with the necessary earthquake data and allow you the option of cacheing your data and the option of accessing data you cached.
This is what I have so far of the curation part of the project. We still need to figure our how to deal with the version column of the dataframe. I'm not quite sure how we should update the version numbers so that they reflect which version of the code we are working on.